requirements.txt
pyserial
opencv-python
numpy 
PyQt5
matplotlib